requirements.txt
requests
beautifulsoup4
google-generativeai
python-dotenv
playwright
matplotlib
crawl4ai
pandas
